add lookup for app/_layout.tsx file
